Access Data From the Global Plant Phenology Data Portal

Access data from the global plant phenology data portal (PPO data portal)

ppo_data(
  genus = NULL,
  specificEpithet = NULL,
  termID = NULL,
  fromYear = NULL,
  toYear = NULL,
  fromDay = NULL,
  toDay = NULL,
  bbox = NULL,
  limit = NULL,
  timeLimit = 60
)

Arguments

genus
(string) a plant genus name

specificEpithet
(string) a plant specific epithet

termID
(string) A single termID from the plant phenology ontology. See the ppo_terms function for more information.

fromYear
(integer) return data from the specified year

toYear
(integer) return data up to and including the specified year

fromDay
(integer) return data starting from the specified day

toDay
(integer) return data up to and including the specified day

bbox
(string) return data within a bounding box. Format is lat,long,lat,long and is structured as a string. Use this website: http://boundingbox.klokantech.com/ to quickly grab a bbox (set format on bottom left to csv and be sure to switch the order from long, lat, long, lat to lat, long, lat, long).

limit
(integer) limit returned data to a specified number of records

timeLimit
(integer) set the limit ofthe amount of time to wait for a response

Value

Return value containing a list with the following components:

`data`: A data frame containing data

`readme`: A string with information about the return package

`citation`: A string with citation information

`number_possible`: An integer with total possible results

`status_code`: An integer with status code returned from server

Details

The ppo_data function returns a list containing the following information: a readme file, citation information, a data frame with data, an integer with the number of records returned and a status code. The function is called with parameters that correspond to values contained in the data itself which act as a filter on the returned record set.

Examples

r1 <- ppo_data(genus = "Quercus", termID='obo:PPO_0002313', limit=10, timeLimit = 4)

#> sending request for data ...

#> https://www.plantphenology.org/api/v2/download/?q=%2Bgenus:Quercus+AND+%2BtermID:"obo:PPO_0002313"+AND+source:USA-NPN,NEON&source=latitude,longitude,year,dayOfYear,termID&limit=10

r2 <- ppo_data(fromDay = 1, toDay = 100, bbox="37,-120,38,-119", limit=10, timeLimit = 4)

#> sending request for data ...

#> https://www.plantphenology.org/api/v2/download/?q=%2Blatitude%3A%3E%3D37+AND+%2Blatitude%3A%3C%3D38+AND+%2Blongitude%3A%3E%3D-120+AND+%2Blongitude%3A%3C%3D-119+AND+%2BdayOfYear:>=1+AND+%2BdayOfYear:<=100+AND+source:USA-NPN,NEON&source=latitude,longitude,year,dayOfYear,termID&limit=10

my_data_frame <- r2$data